requirements.txt
prefect>=2.0.0
clickhouse-driver>=0.2.0
boto3>=1.26.0
requests>=2.28.0
python-telegram-bot>=20.0
pydantic>=2.0.0
python-dotenv>=0.21.0
pandas>=1.5.0